agents/jobdescription/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
colorama==0.4.6
distro==1.9.0
dnspython==2.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
openai==1.107.3
pydantic==2.11.9
pydantic_core==2.33.2
pymongo==4.15.1
PyMuPDF==1.26.4
python-dotenv==1.1.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
watchdog==6.0.0
langgraph
